Fix RPM build process
When doing an RPM build you typically run Makefile.PL in the %build section. This is before whatever you have configured for your var_path has been created if on a fresh machine. When loading the thruk configuration it's final steps are to check if that directory exists, gather user information from it and switch users. This is not necessary during our build process.