Add intelligent page titles
This makes intelligent page titles an option as well as using bookmark names as page titles an option as well.

Removed my last attempt at page titles in favor of this approach which centralizes things a bit and makes everything optional.

Finally, updated FAQ to explain how adding title= to a url on most pages allows for the user to set custom titles. Added this
ability to more pages.