Improved perf data bars to handle more types of perfdata
Updated the logic that builds the perfdata bars to be able to handle almost all of the thresholds specified in the Nagios Plugins Development Guidlines. The only thing these changes don't support is when the '@' symbol is used to do inclusive thresholds. Another thing to note is that the regex to pull apart data was not only modified to handle these other thresholds but to also allow the UOM to be any number of characters long. Previously, this was limited to only allowing two characters however there isn't anything in the guidelines that talks about the max length the UOM can be.