Make set_custom_vars() more extensible
The set_custom_vars function was only able to process 'show_custom_vars' but there will likely be a need in the future to do the same processing on other custom variables that are either specified by the user and the configuration file or by the inner workings of Thruk itself. This will hopefully reduce code and duplication of effort. Also, moved the expansion of macros into the set_custom_vars function as well.