Better sorting of comments and downtimes
This fixes a few things mainly adding the ability to sort the comments
and downtimes on the host and service pages. There were also a few style
changes I made as I saw fit along the way.

Mostly, the style changes were just wording tweaks, capitalization and
formatting. Added output to show current sorting method by default. This
may not be desired but I liked it so I kept it. Can remove it if people
do not like it.