Fixes to Exfoliation theme
Fix typo I made when fixing the themes CSS readability. After fixing the
padding so that it got applied to most browsers I found some of it to be
ugly and prefer it didn't exist and look the way it did before when it was
broken.

Some style fixes around how the pnp4nagios graph image tables and their
borders are displayed.