Add ssi contrib file to turn tickets into links
This is a sample ssi file that can be used as an example on how you can
transform ticket numbers (or whatever regex you want) into a link to an
internal ticketing system.

This works on the host/service information pages as well as the
comments/downtimes pages. It has been tested in Chrome, Firefox, Safari
and IE.