Add support for recurring downtimes page
The recurring downtime page is much trickier than the other pages
because an extra column will be added depending on whether a host or
service is displayed.

The way around this problem is instead of counting from the first child
forward from the column index that was found, count backward from the
end where no extra rows will be added.

This works in Chrome, Firefox and Safari consistantly. There are some
issues with IE. Support for 'nth-last-child' came in IE9. Tests showed
this works as long as IE9 or greater is in use. However, Thruk currently
sends a X-UA-Compatible header specifying IE8 which breaks this unless
overridden by the user.