Fix reschedule check forced onto next line
In many of the themes the "Re-schedule next check of this service" was
getting pushed onto the next line. In most cases just shortening the
length of the string seemed to do the trick. Removing the space between
'Now' and the parenthesis seemed appropriate. There are other links in
Thruk (the link box is a good example) where links are encapsulated in
parenthesis and there is no space seperating them.

In the case of the Neat theme some additional CSS changes were needed.