Add ability to generate URL for excel export
Create new button and input box to help create a URL for an excel export
that includes the selected columns.

Previously, one would have to look at their web server logs for the URL.

First attempt was to try and do this by opening the export in a new
winodow/tab but it seems most browsers (Chrome and FF atleast) will
cloes that tab immediately after the excel sheet was downloaded.

This lead to the URL itself being generated by the user once they are
happy with the options they have selected for their export.