Add 'expand_user_macros' option
With the addition of the new action menu capabilities to Thruk there is
an added need to allow $USERx$ macros to be expanded while still
limiting their exposure when a host/service command line or custom
variable is expanded.

The 'expand_user_macros' option allows for granular control over which
$USERx$ macros are expanded when the command line and custom variables
are expanded while still allowing server actions the ability to use any
$USERx$ macro.

This new option also still retains the ability of completely turn off
the expansion of all $USERx$ macros by just not specifying anything in
the configuration. To turn expansion on for all macros simply
specifiying 'ALL' will do the trick.

The comparison between the $USERx$ macros and the values supplied in
'expand_user_macros' is not currently capable of handling regular
expressions. This seemed a bit overkill but if there is a good value add
then it can certainly be changed to do so in the future.

Documentation was updated to reflect all of the changes including a few
clarifying changes to the action menu peices that show $USERx$ macros
are now supported and how to specify multiple arguments to server
actions.