Add ability to display action menu title
The action menu can get lost on the extinfo pages. This can obviously
depend on the icon used as well but having the menu title displayed
makes the menu a bit more visible. If someone doesn't like the menu
title displayed they can just not use one in their configuration.