reports: include any ok event always in availability calculation
Somehow nagios can change from a Hard Critical into a Soft Critical which then results in a soft ok.
Any ok state always resets the current problem, so no matter if a ok is soft or hard, we have
to count it in. Otherwise we could end up with a critical last entry in the logfile, even if
the current state is ok.