Data Import Enhancements
It fixes the problem from https://github.com/sni/Thruk/issues/694 as MariaDB reports errors if the data does not match.
In addition on my system state_type reported more state_types. The list might not be complete. Check the log files for more types.
Changed state_type to undef, because $dbh->quote($state_type) will change NULL to 'NULL', which is not allowed in the INSERT command.
Changed the SQL error logging for easier troubleshooting.