Introduces a configurable switch to globally activate/disactivate multiple backend sites
Introduces the parameter ‘site_global_switch’ to the thruk.conf file.
Setting the parameter to 1,  activates two icons on the GUI which allows
all configured backend sites to be activated or deactivated with one
click.

The change provides similar functionality to the ‘compact site panel’.
However the solution differs in that the list of configured backend
sites always remains visible, where as on the ‘compact site panel’ ,
only provides a summary description of the total number of active or
disactive sites.