Global backend activation switch made a function of number of backends
The point at which the global backend activation switch becomes visibile
is defined by the value of the 'site_global_switch'. The switch is
rendered if the number of backends is greater or equal to the
number specified by the parameter. The switch is hidden if the number of
backends is less than the specified value or if the parameter is set to
zero.