panorama: reset state for non-state icon
sometimes it happens that non-state icons (ex.: static images) have a state in the runtime file. This might happen
if the number of icon changes in dynamic generated dashboards. This leads to changed z-index because of the raised
z-index for non-ok icons. This then might lead to background images overlapping stateful icons.